# src/main/resources/api-specs/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cosmocat Marketplace API",
    "description": "This API exposes endpoints to manage products in the Cosmocat Marketplace.",
    "contact": {
      "name": "Cosmocat Marketplace Team",
      "email": "cosmocat@marketplace.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://choosealicense.com/licenses/mit/"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Server URL in Development environment"
    }
  ],
  "tags": [
    {
      "name": "Product Management",
      "description": "APIs for managing products in the marketplace"
    }
  ],
  "paths": {
    "/api/v1/products/{id}": {
      "get": {
        "tags": [
          "Product Management"
        ],
        "summary": "Get product by ID",
        "description": "Retrieve a specific product by its unique identifier",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CosmoApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CosmoApiResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Product Management"
        ],
        "summary": "Update product",
        "description": "Update an existing product's information",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CosmoApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CosmoApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CosmoApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product Management"
        ],
        "summary": "Delete product",
        "description": "Delete a product from the marketplace",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CosmoApiResponseVoid"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CosmoApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products": {
      "get": {
        "tags": [
          "Product Management"
        ],
        "summary": "Get all products",
        "description": "Retrieve a list of all available products in the marketplace",
        "operationId": "getAllProducts",
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CosmoApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product Management"
        ],
        "summary": "Create new product",
        "description": "Create a new product in the marketplace",
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid request data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CosmoApiResponse"
                }
              }
            }
          },
          "201": {
            "description": "Product created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CosmoApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/search": {
      "get": {
        "tags": [
          "Product Management"
        ],
        "summary": "Search products by name",
        "description": "Search for products by name (case-insensitive partial match)",
        "operationId": "searchProducts",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Product name to search for",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "laptop"
          }
        ],
        "responses": {
          "200": {
            "description": "Search completed successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CosmoApiResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProductUpdateRequest": {
        "type": "object",
        "description": "Request object for updating an existing product",
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Stellar Bluetooth Headphones Pro",
            "maxLength": 100,
            "minLength": 2
          },
          "description": {
            "type": "string",
            "description": "Product description",
            "example": "Premium wireless headphones with enhanced noise cancellation",
            "maxLength": 1000,
            "minLength": 0
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Product price",
            "example": 149.99,
            "maximum": 999999.99,
            "minimum": 0.01
          },
          "currency": {
            "type": "string",
            "description": "Currency code",
            "example": "USD",
            "pattern": "^[A-Z]{3}$"
          },
          "stockQuantity": {
            "type": "integer",
            "format": "int32",
            "description": "Available stock quantity",
            "example": 30,
            "maximum": 100000,
            "minimum": 0
          },
          "sku": {
            "type": "string",
            "description": "Stock Keeping Unit",
            "example": "GALAXY-WBH-PRO-001",
            "maxLength": 50,
            "minLength": 3,
            "pattern": "^[A-Z0-9-]+$"
          },
          "image": {
            "type": "string",
            "description": "Product image URL",
            "example": "https://example.com/images/headphones-pro.jpg",
            "pattern": "^(https?://.*|)$"
          },
          "weight": {
            "type": "number",
            "format": "double",
            "description": "Product weight in kg",
            "example": 0.3,
            "maximum": 1000.0,
            "minimum": 0.001
          },
          "dimensions": {
            "type": "string",
            "description": "Product dimensions",
            "example": "22x20x9 cm",
            "maxLength": 100,
            "minLength": 0,
            "pattern": "^\\d+(\\.\\d+)?\\s*x\\s*\\d+(\\.\\d+)?\\s*x\\s*\\d+(\\.\\d+)?\\s*(cm|mm|m|inch|in)$"
          },
          "availabilityStatus": {
            "type": "string",
            "description": "Product availability status",
            "enum": [
              "AVAILABLE",
              "OUT_OF_STOCK",
              "DISCONTINUED",
              "PRE_ORDER"
            ],
            "example": "AVAILABLE"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "description": "Category ID",
            "example": 1
          }
        }
      },
      "CosmoApiResponse": {
        "type": "object",
        "description": "Standard API response wrapper",
        "properties": {
          "status": {
            "type": "string",
            "description": "Response status",
            "example": "success"
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Operation completed successfully"
          },
          "data": {
            "description": "Response data"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Response timestamp"
          }
        }
      },
      "ProductCreateRequest": {
        "type": "object",
        "description": "Request object for creating a new product",
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Stellar Bluetooth Headphones",
            "maxLength": 100,
            "minLength": 2
          },
          "description": {
            "type": "string",
            "description": "Product description",
            "example": "High-quality cosmic headphones with stellar noise cancellation technology",
            "maxLength": 1000,
            "minLength": 0
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Product price",
            "example": 129.99,
            "maximum": 999999.99,
            "minimum": 0.01
          },
          "currency": {
            "type": "string",
            "description": "Currency code",
            "example": "USD",
            "pattern": "^[A-Z]{3}$"
          },
          "stockQuantity": {
            "type": "integer",
            "format": "int32",
            "description": "Available stock quantity",
            "example": 50,
            "maximum": 100000,
            "minimum": 0
          },
          "sku": {
            "type": "string",
            "description": "Stock Keeping Unit",
            "example": "STAR-WBH-001",
            "maxLength": 50,
            "minLength": 3,
            "pattern": "^[A-Z0-9-]+$"
          },
          "image": {
            "type": "string",
            "description": "Product image URL",
            "example": "https://example.com/images/headphones.jpg",
            "pattern": "^(https?://.*|)$"
          },
          "weight": {
            "type": "number",
            "format": "double",
            "description": "Product weight in kg",
            "example": 0.25,
            "maximum": 1000.0,
            "minimum": 0.001
          },
          "dimensions": {
            "type": "string",
            "description": "Product dimensions",
            "example": "20x18x8 cm",
            "maxLength": 100,
            "minLength": 0,
            "pattern": "^\\d+(\\.\\d+)?\\s*x\\s*\\d+(\\.\\d+)?\\s*x\\s*\\d+(\\.\\d+)?\\s*(cm|mm|m|inch|in)$"
          },
          "availabilityStatus": {
            "type": "string",
            "description": "Product availability status",
            "enum": [
              "AVAILABLE",
              "OUT_OF_STOCK",
              "DISCONTINUED",
              "PRE_ORDER"
            ],
            "example": "AVAILABLE"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "description": "Category ID",
            "example": 1
          }
        },
        "required": [
          "availabilityStatus",
          "currency",
          "name",
          "price",
          "stockQuantity"
        ]
      },
      "CosmoApiResponseVoid": {
        "type": "object",
        "description": "Standard API response wrapper",
        "properties": {
          "status": {
            "type": "string",
            "description": "Response status",
            "example": "success"
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Operation completed successfully"
          },
          "data": {
            "type": "object",
            "description": "Response data"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Response timestamp"
          }
        }
      }
    }
  }
}